Refactor tests inside module to remove need for copying
