+++ packages/api-openfin-demo/package.json
@@ -7,9 +7,9 @@
     "serve": "live-server --port=5000 --no-browser src/",
     "launch": "openfin --launch --config ./src/app.json",
     "start": "npm-run-all --parallel serve launch",
-    "postinstall": "rimraf src && rimraf test && copyfiles -f \"node_modules/containerjs-api-specification/src/**\" src && copyfiles -f \"node_modules/containerjs-api-openfin/dist/**\" src && copyfiles -f \"node_modules/containerjs-api-tests/test/**\" test",
+    "postinstall": "rimraf src && copyfiles -f \"node_modules/containerjs-api-specification/src/**\" src && copyfiles -f \"node_modules/containerjs-api-openfin/dist/**\" src",
     "test": "exit 0",
-    "test:ui": "mocha test/*.js"
+    "test:ui": "mocha test/main.spec.js"
   },
   "author": "",
   "license": "Apache-2.0",
